Ravage script under the enemy directory
One thing to notice here is about the NavMesh. I use the NaveMesh for the navigation of the enemy. At the same time, we need to bake the plane or terrain where the enemy can walk on. For doing that, we have to change those scenes to static. Not quite sure whether it will cause any problems.